+++ test/otel/exporter/chdb_attribute_projection_test.clj
@@ -36,7 +36,8 @@
      (installer/install-approved!
       (backend/memory-backend) manifest
       {:target target
-       :observe-columns #(into {} @observed)
+       :observe-columns #(vector {:columns (into {} @observed)
+                                  :signal :spans :table "otel_traces"})
        :execute-ddl!
        (fn [_]
          (let [{:keys [name type]} (nth columns @next-column)]